requirements.txt
matplotlib==3.9.3
numpy==2.1.3
opencv_python==4.10.0.84
Pillow==11.0.0
PyAutoGUI==0.9.54
pytesseract==0.3.13
